Move dump and pack to tlvc_text and add more tests
